Improve tests and "convert" method
Before these changes, the result of convert method with same parameter changed everytime if we used the same object to call this function. The result of convert with same parameter was constant if we used different objects. To avoid this issue, we were forced to create different objects but that's not how it should be done.

Now, the program clears the list of used tags in order to always get an empty list when we call this method.