feat: Migrate from Java 11 to 17
The main reason is the cross platform desktop app (running on Windows, Mac and Linux) we will build using JavaFX. It's possible to use other versions inferior than 17 but we will miss good documentation. We thought about using only version 17 for the desktop module only then we decided to use it for all modules.