fix: compilation failure due to  4 modules (dependencies of eu.hansolo.tilesfx)  not found
There was an issue while trying to run javafx:run. In fact, dependencies of eu.hansolo.tilesfx ( countries, heatmap, toolbox,  toolboxfx ) were not found by maven 3.8.6. This is why this dependency tilesfx was commented.  One way to solve this problem is  to add these dependencies in the pom while using maven 3.8.6. We were supposed to use that solution but after starting using maven 4, the problem is finally solved without adding these 4 dependencies.