fix: uncommenting-module eu.hansolo.tilesfxtilesfx
Previously, we thought that moving from Maven 3.8.6 to 4 helps us to not be forced to add the dependencies needed by tilesfx to run the desktop app but it's wrong. It looked like it was the solution because the module tilesfx was commented in module-info.java. Module tilesfx was commented in order to solve the issue later. Now, we are happy that it is finally solved.